deleted:    Astah/projLingProg7.asta 	deleted:    Astah/projLingProg7Controller.puml 	new file:   Astah/projLingProg8.asta 	renamed:    Astah/projLingProg7CasosdeUso.puml -> Astah/projLingProg8CasosdeUso.puml 	renamed:    Astah/projLingProg7Classes.puml -> Astah/projLingProg8Classes.puml 	modified:   Database/estoque.db 	deleted:    t1/pom.xml 	deleted:    t1/src/main/java/br/com/restaurante/sistema_estoque_restaurante/App.java 	deleted:    t1/src/main/java/br/com/restaurante/sistema_estoque_restaurante/dao/dbcon/AcessoSQLite.java 	deleted:    t1/src/main/java/br/com/restaurante/sistema_estoque_restaurante/dao/dbcon/IngredienteDao.java 	deleted:    t1/src/main/java/br/com/restaurante/sistema_estoque_restaurante/dao/dbcon/MovimentacaoDao.java 	deleted:    t1/src/main/java/br/com/restaurante/sistema_estoque_restaurante/dao/dbcon/ProdutoDao.java 	deleted:    t1/src/main/java/br/com/restaurante/sistema_estoque_restaurante/dao/dbcon/ReceitaDao.java 	deleted:    t1/src/main/java/br/com/restaurante/sistema_estoque_restaurante/dao/dbcon/UnidadeDao.java 	deleted:    t1/src/main/java/br/com/restaurante/sistema_estoque_restaurante/dao/dbcon/UserDao.java 	deleted:    t1/src/main/java/br/com/restaurante/sistema_estoque_restaurante/model/Cardapio.java 	deleted:    t1/src/main/java/br/com/restaurante/sistema_estoque_restaurante/model/Ingrediente.java 	deleted:    t1/src/main/java/br/com/restaurante/sistema_estoque_restaurante/model/MovimentacaoEstoqueDeProduto.java 	deleted:    t1/src/main/java/br/com/restaurante/sistema_estoque_restaurante/model/Produto.java 	deleted:    t1/src/main/java/br/com/restaurante/sistema_estoque_restaurante/model/Receita.java 	deleted:    t1/src/main/java/br/com/restaurante/sistema_estoque_restaurante/model/TipoCadastro.java 	deleted:    t1/src/main/java/br/com/restaurante/sistema_estoque_restaurante/model/TipoMovimentacao.java 	deleted:    t1/src/main/java/br/com/restaurante/sistema_estoque_restaurante/model/Unidade.java 	deleted:    t1/src/main/java/br/com/restaurante/sistema_estoque_restaurante/model/User.java 	deleted:    t1/src/main/java/module-info.java 	deleted:    t1/src/main/resources/META-INF/persistence.xml 	modified:   t2/src/main/java/sdedr/App.java 	modified:   t2/src/main/java/sdedr/ctrl/MovimentacaoEstoqueDeProdutoCtrl.java 	modified:   t2/src/main/java/sdedr/dao/IngredienteDao.java 	modified:   t2/src/main/java/sdedr/dao/MovimentacaoDao.java 	modified:   t2/src/main/java/sdedr/dao/ProdutoDao.java 	modified:   t2/src/main/java/sdedr/dao/dbcon/AcessoSQLite.java 	modified:   t2/src/main/java/sdedr/view/Cadastros/CadastroReceitaView.java 	modified:   t2/src/main/java/sdedr/view/LoginView.java 	modified:   t2/src/main/java/sdedr/view/ManterGeralView.java 	modified:   t2/src/main/java/sdedr/view/MenuTmpView.java 	modified:   t2/src/main/java/sdedr/view/MenuView.java 	modified:   t2/src/main/java/sdedr/view/MovimentacaoEstoqueView.java 	new file:   t2/src/main/java/sdedr/view/ReceitasView.java

+++ t2/src/main/java/sdedr/dao/IngredienteDao.java
@@ -28,7 +28,7 @@ public class IngredienteDao {
   private void formatarIngrediente(Ingrediente resultado, ResultSet saida) throws SQLException {
     ProdutoDao produtoDao = new ProdutoDao();
     Produto produto = new Produto();
-    produtoDao.retornar(saida.getInt("id_p"), produto);
+    produtoDao.retornar(saida.getLong("id_p"), produto);
 
     ReceitaDao receitaDao = new ReceitaDao();
     Receita receita = new Receita();
